Replace .NET WebBrowser with CefSharp.

+++ src/QSP/TOPerfCalculation/Airbus/DataClasses/TableDataNode.cs
@@ -5,7 +5,7 @@ namespace QSP.TOPerfCalculation.Airbus.PerfData
     public class TableDataNode
     {
         public string Flaps { get; set; }
-        public int IsaOffset { get; set; }
+        public double IsaOffset { get; set; }
         public Table2D Table { get; set; }
     }
 }
